Production code included
In this section it is included the scripts needed to make predictions
The config file has all the features and paths to run the scripts
In the preprocessors file are defined all the classes for the pipeline. This classes inherit methods from the sklearn library
In the pipeline file it is defined the pipeline created, with all the preprocessor and the model as the last step of the pipeline
The train_pipeline file loads the dataset and applies the fit method for the pipeline, then save the pipeline
Finally, the make_predictions file loads this trained pipeline and outputs predictions on the test dataset.

In addition to that, jupyter notebooks are moved to a folder and the original train.csv file is renamed as houseprice.csv